# src/config/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
      "version": "1.0.0",
      "title": "Blog Post API",
      "description": "This is Post Share application. You can register, login and share post in system. You can display to posts of users you followed.",
      "contact":{  
         "email":"berkayc78@gmail.com"
      }
  },
  "host": "localhost:8080",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/register": {
        "post": {
            "operationId":"registerUser",
            "tags": ["Authentication"],
            "summary":"Create new user in system",
            "parameters":[
                {
                    "name": "Register User",
                    "in": "body",
                    "description": "User that will be created",
                    "schema": {
                        "type":"Object",
                        "properties":{
                            "User":{
                                "$ref": "#/definitions/UserToRegister"
                            }
                        }
                        
                    },
                    "required":true
            }],
            "responses": {
                "200": {
                    "description": "User was succesfully created but token could not be generated."
                },
                "201": {
                    "description": "User was succesfully created.",
                    "headers":{  
                        "Authorization":{
                            "type":"string",
                            "format":"Bearer",
                            "description":"token of authenticated user with format like 'Bearer xxxxx'"
                        }
                    }
                },
                "400": {
                    "description": "User could not be created.",
                    "schema":{
                      "$ref":"#/definitions/ApiErrorResponse"
                    }   
                }
            }
        }
    },
    "/auth/login": {
        "post": {
            "operationId":"loginUser",
            "tags": ["Authentication"],
            "summary":"Login system and get token in header",
            "parameters":[
                {
                    "name": "Credentials",
                    "in": "body",
                    "description": "Credentials that will be login to system.",
                    "schema": {
                        "type":"Object",
                        "$ref": "#/definitions/Credentials"
                    },
                    "required":true
                }
            ],
            "responses": {
                "201": {
                    "description": "User is logged in.",
                    "headers":{  
                        "Authorization":{
                            "type":"string",
                            "format":"Bearer",
                            "description":"token of authenticated user with format like 'Bearer xxxxx'"
                        }
                    }
                },
                "400": {
                    "description": "User could not be logged in.",
                    "schema":{
                      "$ref":"#/definitions/ApiErrorResponse"
                    }   
                }
                
            }
        }
    },
    "/auth/logout": {
        "get": {
            "operationId":"logoutUser",
            "tags": ["Authentication"],
            "summary":"Logout authenticated user who has token in heeader",
            "produces":[],
            "parameters":[
              {
                "name": "Authorization",
                "in": "header",
                "description": "User token that given with login. (Authorization: Bearer xxxxx )",
                "required":true
              }
            ],
            "responses": {
                "200": {
                    "description": "User was succesfully logged out."
                },
                "401": {
                  "description": "Could not log out. Maybe, User token that sending in header, is not valid!",
                  "schema":{
                    "$ref":"#/definitions/ApiErrorResponse"
                  }   
                }
            }
        }
    },
...
  },
  "definitions": {
...
    "UserToRegister": {
        "required": [
            "Username",
            "Password",
            "Fullname"
        ],
        "properties": {
            "Username": {
                "type": "string"
            },
            "Password": {
                "type": "string"
            },
            "Fullname":{
                "type": "string"
            },
            "Contact": {
                "required": [
                    "Email"
                ],
                "properties": {
                    "Email": {
                        "type": "string"
                    },
                    "Phone": {
                        "type": "string"
                    }
                }
            }
        }
    },
...
    "Credentials":{
        "required":[
            "Username",
            "Password"
        ],
        "properties":{
            "Username": {
                "type": "string"
            },
            "Password": {
                "type":"string"
            }
        }                   
    },
    "ApiErrorResponse": {
        "properties":{
          "error":{
            "type":"string"
          }
        }
    },
...
  }
}